Bug quashing session done, and redundent tests removed

+++ src/graph.h
@@ -9,7 +9,7 @@ struct Node
     char Name[256];
     char Path[512];
     bool IsDirectory;
-    float Mass;
+    uint64_t Mass;
     float Radius;
     int Depth;
     Vector3 Position;
